requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1

langchain==0.2.7
langchain-core==0.2.22
langchain-community==0.2.7

langchain-mongodb==0.1.6
langchain-text-splitters==0.2.2

langchain-openai==0.1.17
langchain-cohere==0.1.9
langchain-fireworks==0.1.5
langchain-google-genai==1.0.7
langchain-groq==0.1.6
langchain-ollama==0.1.0

rich==13.7.1
pandas==2.2.2
tqdm==4.66.4
openpyxl==3.1.5
pymongo==4.8.0

llama-index-core==0.10.55
llama-parse==0.4.9

python-dotenv==1.0.1
streamlit==1.36.0

azure-storage-blob==12.22.0
azure-identity==1.17.1
adlfs==2024.7.0
